added router and pages folder
