+++ src/Components/Header.css
@@ -1,0 +1,20 @@
+header {
+    background-color: whitesmoke;
+}
+
+header ul {
+    display: flex;
+    align-items: center;
+    margin-top: 0;
+    padding-left: 10px;
+}
+
+header ul li {
+    margin-right: 25px;
+    list-style: none;
+}
+
+header li a {
+    text-decoration: none;
+    font-size: 20px;
+}
