+++ src/Components/Header.js
@@ -1,6 +1,8 @@
 import React from "react";
+import {Link} from "react-router-dom";
+import "./Header.css"
 
-const Header = () => {
+function Header(){
   return (
     <header
       style={{
@@ -10,9 +12,21 @@ const Header = () => {
         padding: "10px 0px 0px 10px",
       }}
     >
-      <h2 style={{ color: "navy", textShadow: "2px 4px 5px gray" }}>
-        reactcdv
-      </h2>
+    <ul>
+      <li>
+        <Link to="/" >      
+          <h2 style={{ color: "navy", textShadow: "2px 4px 5px gray" }}>
+            REACTCDV
+          </h2>
+        </Link>
+      </li>
+      <li>
+        <Link to="/Tab1">
+          TAB1
+        </Link>
+      </li>
+    </ul>
+
     </header>
   );
 };
