+++ src/Pages/Tab.js
@@ -1,0 +1,9 @@
+import React from "react"
+
+function Tab(){
+    return (
+        <h1>Tab</h1>
+    )
+}
+
+export default Tab;
